+++ priv/repo/migrations/20211115173048_fill_vf_claim.exs
@@ -26,9 +26,9 @@ def change() do
 		add :resource_classified_as, {:array, :text}
 		add :resource_conforms_to_id, references("vf_resource_specification")
 		add :resource_quantity_has_unit_id, references("vf_unit")
-		add :resource_quantity_has_numerical_value, :float
+		add :resource_quantity_has_numerical_value, :decimal
 		add :effort_quantity_has_unit_id, references("vf_unit")
-		add :effort_quantity_has_numerical_value, :float
+		add :effort_quantity_has_numerical_value, :decimal
 		add :triggered_by_id, references("vf_economic_event")
 		add :due, :timestamptz
 		add :created, :timestamptz
