+++ priv/repo/migrations/20211115181352_fill_vf_settlement.exs
@@ -23,9 +23,9 @@ def change() do
 		add :settled_by_id, references("vf_economic_event"), null: false
 		add :settles_id, references("vf_claim"), null: false
 		add :resource_quantity_has_unit_id, references("vf_unit")
-		add :resource_quantity_has_numerical_value, :float
+		add :resource_quantity_has_numerical_value, :decimal
 		add :effort_quantity_has_unit_id, references("vf_unit")
-		add :effort_quantity_has_numerical_value, :float
+		add :effort_quantity_has_numerical_value, :decimal
 		add :note, :text
 		timestamps()
 	end
